Rename torch source file and update build note
