+++ AGENTS.md
@@ -1,0 +1,5 @@
+# Agent Instructions
+
+- This repository is focused exclusively on the torch example located in `examples/torch`.
+- When validating changes, building the `llama-torch-example` target is sufficient; no additional tests are required.
+- Use `-j 8` when invoking `cmake --build` so the build runs with reasonable parallelism.
