+++ examples/torch/main.cpp
@@ -1,0 +1,8 @@
+#include "torch.h"
+
+#include <iostream>
+
+int main() {
+    std::cout << "hello world" << std::endl;
+    return 0;
+}
