+++ src/CMakeLists.txt
@@ -35,7 +35,6 @@ add_library(llama
             unicode-data.cpp
             unicode.cpp
             unicode.h
-            ggml-torch.cpp
             )
 
 target_include_directories(llama PRIVATE .)
